requirements.txt
streamlit
torch
einops
timm
transformers
Pillow
requests
matplotlib
